# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.userEnterData:
    properties:
      email:
        type: string
      lastPassword:
        type: string
      password:
        type: string
    type: object
...
host: localhost:1111
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: An e-commerce site in Go using Gin framework
  title: E-commerce
  version: "1.0"
paths:
...
  /user/changepassword:
    patch:
      consumes:
      - application/json
      description: Change user password
      parameters:
      - description: Password change data
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/controllers.userEnterData'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerToken: []
      summary: Change Password
      tags:
      - Authentication
      - Users
...
schemes:
- http
swagger: "2.0"